+++ lib/karafka/pro/processing/strategies/default.rb
@@ -54,8 +54,11 @@ module Karafka
           #   already processed but rather at the next one. This applies to both sync and async
           #   versions of this method.
           def mark_as_consumed(message, offset_metadata = @_current_offset_metadata)
+            # If we are inside a transaction than we can just mark as consumed within it
             if @_in_transaction
               mark_in_transaction(message, offset_metadata, true)
+            elsif @_in_transaction_marked
+              mark_in_memory(message)
             else
               # seek offset can be nil only in case `#seek` was invoked with offset reset request
               # In case like this we ignore marking
@@ -64,16 +67,15 @@ module Karafka
               return true if coordinator.seek_offset > message.offset
               return false if revoked?
 
-              # If we have already marked this successfully in a transaction that was running
-              # we should not mark it again with the client offset delegation but instead we should
-              # just align the in-memory state
-              if @_in_transaction_marked
-                coordinator.seek_offset = message.offset + 1
+              # If we are not inside a transaction but this is a transactional topic, we mark with
+              # artificially created transaction
+              stored = if producer.transactional?
+                         mark_with_transaction(message, offset_metadata, true)
+                       else
+                         client.mark_as_consumed(message, offset_metadata)
+                       end
 
-                return true
-              end
-
-              return revoked? unless client.mark_as_consumed(message, offset_metadata)
+              return revoked? unless stored
 
               coordinator.seek_offset = message.offset + 1
             end
@@ -92,6 +94,8 @@ module Karafka
           def mark_as_consumed!(message, offset_metadata = @_current_offset_metadata)
             if @_in_transaction
               mark_in_transaction(message, offset_metadata, false)
+            elsif @_in_transaction_marked
+              mark_in_memory(message)
             else
               # seek offset can be nil only in case `#seek` was invoked with offset reset request
               # In case like this we ignore marking
@@ -100,13 +104,15 @@ module Karafka
               return true if coordinator.seek_offset > message.offset
               return false if revoked?
 
-              if @_in_transaction_marked
-                coordinator.seek_offset = message.offset + 1
+              # If we are not inside a transaction but this is a transactional topic, we mark with
+              # artificially created transaction
+              stored = if producer.transactional?
+                         mark_with_transaction(message, offset_metadata, false)
+                       else
+                         client.mark_as_consumed!(message, offset_metadata)
+                       end
 
-                return true
-              end
-
-              return revoked? unless client.mark_as_consumed!(message, offset_metadata)
+              return revoked? unless stored
 
               coordinator.seek_offset = message.offset + 1
             end
@@ -128,7 +134,7 @@ module Karafka
           #   managing multiple producers. If not provided, default producer taken from `#producer`
           #   will be used.
           #
-          # @param block [Proc] code that we want to run in a transaction
+          # @yield code that we want to run in a transaction
           #
           # @note Please note, that if you provide the producer, it will reassign the producer of
           #   the consumer for the transaction time. This means, that in case you would even
@@ -136,7 +142,7 @@ module Karafka
           #   reassigned producer and not the initially used/assigned producer. It is done that
           #   way, so the message producing aliases operate from within transactions and since the
           #   producer in transaction is locked, it will prevent other threads from using it.
-          def transaction(active_producer = producer, &block)
+          def transaction(active_producer = producer)
             default_producer = producer
             self.producer = active_producer
 
@@ -150,7 +156,14 @@ module Karafka
             @_in_transaction = true
             @_in_transaction_marked = false
 
-            producer.transaction(&block)
+            producer.transaction do
+              yield
+
+              # Ensure this transaction is rolled back if we have lost the ownership of this
+              # transaction. We do it only for transactions that contain offset management as for
+              # producer only, this is not relevant.
+              raise Errors::AssignmentLostError if @_in_transaction_marked && revoked?
+            end
 
             @_in_transaction = false
 
@@ -174,6 +187,8 @@ module Karafka
             @_transaction_marked.each do |marking|
               marking.pop ? mark_as_consumed(*marking) : mark_as_consumed!(*marking)
             end
+
+            true
           ensure
             self.producer = default_producer
 
@@ -204,6 +219,52 @@ module Karafka
             @_in_transaction_marked = true
             @_transaction_marked ||= []
             @_transaction_marked << [message, offset_metadata, async]
+          end
+
+          # @private
+          # @param message [Messages::Message] message we want to commit inside of a transaction
+          # @param offset_metadata [String, nil] offset metadata or nil if none
+          # @param async [Boolean] should we mark in async or sync way (applicable only to post
+          #   transaction state synchronization usage as within transaction it is always sync)
+          # @return [Boolean] false if marking failed otherwise true
+          def mark_with_transaction(message, offset_metadata, async)
+            # This flag is used by VPs to differentiate between user initiated transactions and
+            # post-execution system transactions.
+            @_transaction_internal = true
+
+            transaction do
+              mark_in_transaction(message, offset_metadata, async)
+            end
+
+            true
+          # We handle both cases here because this is a private API for internal usage and we want
+          # the post-user code execution marking with transactional producer to result in a
+          # boolean state of marking for further framework flow. This is a normalization to make it
+          # behave the same way as it would behave with a non-transactional one
+          rescue ::Rdkafka::RdkafkaError, Errors::AssignmentLostError
+            false
+          ensure
+            @_transaction_internal = false
+          end
+
+          # Marks the current state only in memory as the offset marking has already happened
+          # using the producer transaction
+          # @param message [Messages::Message] last successfully processed message.
+          # @return [Boolean] true if all good, false if we lost assignment and no point in marking
+          def mark_in_memory(message)
+            # seek offset can be nil only in case `#seek` was invoked with offset reset request
+            # In case like this we ignore marking
+            return true if coordinator.seek_offset.nil?
+            # Ignore earlier offsets than the one we already committed
+            return true if coordinator.seek_offset > message.offset
+            return false if revoked?
+
+            # If we have already marked this successfully in a transaction that was running
+            # we should not mark it again with the client offset delegation but instead we should
+            # just align the in-memory state
+            coordinator.seek_offset = message.offset + 1
+
+            true
           end
 
           # No actions needed for the standard flow here